task_8/requirements.txt
Django==2.0rc1
zappa
zappa-django-utils
djangorestframework
django-storages
django-webpack-loader